requirements.txt
opencv_contrib_python==4.5.2.52
Flask==1.1.2
matplotlib==3.1.2
numpy==1.17.4
streamlit==0.88.0
pandas==1.2.4
Pillow==8.3.2
scikit_learn==0.24.2
